+++ packages/@tinacms/toolkit/src/packages/react-toolbar/components/DesktopLabel.tsx
@@ -20,7 +20,6 @@ import styled from 'styled-components'
 
 export const DesktopLabel = styled.span`
   all: unset;
-  color: inherit;
   display: none;
   @media (min-width: 1030px) {
     display: inline;