+++ packages/tinacms/src/admin/pages/CollectionCreatePage.tsx
@@ -12,37 +12,33 @@ limitations under the License.
 */
 
 import React from 'react'
-import { Form, FormBuilder } from '@tinacms/toolkit'
+import { Form, FullscreenFormBuilder } from '@tinacms/toolkit'
 import { useParams, useHistory, Link } from 'react-router-dom'
 
-import type { TinaCMS } from '@tinacms/toolkit'
-
-import useGetDocumentFields from '../hooks/useGetDocumentFields'
-
 import GetCMS from '../components/GetCMS'
+import GetDocumentFields from '../components/GetDocumentFields'
 
-const GetDocumentFields = ({
-  cms,
-  collectionName,
-  children,
-}: {
-  cms: TinaCMS
-  collectionName: string
-  children: any
-}) => {
-  const { collection, fields } = useGetDocumentFields(cms, collectionName)
-  if (!collection || !fields) {
-    return null
-  }
-  return <>{children(collection, fields)}</>
-}
+import type { TinaCMS } from '@tinacms/toolkit'
+import { transformDocumentIntoMutationRequestPayload } from '../../hooks/use-graphql-forms'
 
 const createDocument = async (
   cms: TinaCMS,
   collection: { name: string },
+  mutationInfo: { includeCollection: boolean; includeTemplate: boolean },
   values: any
 ) => {
-  const { relativePath, ...params } = values
+  const { relativePath, ...leftover } = values
+  const { includeCollection, includeTemplate } = mutationInfo
+  const params = transformDocumentIntoMutationRequestPayload(
+    {
+      _collection: collection.name,
+      ...leftover,
+    },
+    {
+      includeCollection,
+      includeTemplate,
+    }
+  )
 
   await cms.api.tina.request(
     `mutation($collection: String!, $relativePath: String!, $params: DocumentMutation!) {
@@ -56,7 +52,7 @@ const createDocument = async (
       variables: {
         collection: collection.name,
         relativePath,
-        params: { [collection.name]: { ...params } },
+        params,
       },
     }
   )
@@ -70,7 +66,7 @@ const CollectionCreatePage = () => {
     <GetCMS>
       {(cms: TinaCMS) => (
         <GetDocumentFields cms={cms} collectionName={collectionName}>
-          {(collection, fields) => {
+          {(collection, fields, mutationInfo) => {
             const form = new Form({
               id: 'create-form',
               label: 'form',
@@ -87,23 +83,19 @@ const CollectionCreatePage = () => {
                 ...fields,
               ],
               onSubmit: async (values) => {
-                await createDocument(cms, collection, values)
+                await createDocument(cms, collection, mutationInfo, values)
                 history.push(`/admin/collections/${collection.name}`)
               },
             })
             return (
-              <>
-                <h3 className="text-xl mb-6">
-                  <Link
-                    className="opacity-80 hover:opacity-100 transition-opacity ease-out"
-                    to={`/admin/collections/${collection.name}`}
-                  >
-                    {collection.label}
-                  </Link>{' '}
-                  - Create New
-                </h3>
-                <FormBuilder form={form} />
-              </>
+              <div className="w-full h-screen">
+                <div className="flex flex-col items-center w-full flex-1">
+                  <FullscreenFormBuilder
+                    label={collection.label + ' - Create New'}
+                    form={form}
+                  />
+                </div>
+              </div>
             )
           }}
         </GetDocumentFields>